api: Fix operator precedence and conversion
Problem:
  - commit 89088fe37d9a21c1fd7c6c372e3f91b8acaf8b2c contained a wrong
    fix for the second problem with operator precedence as well as an
    issue regarding implicit type conversion

Solution:
  - Fix the parentheses to realise the desired operator precedence
  - Explicitly cast to avoid a conversion warning

+++ src/nvim/api/buffer.c
@@ -174,7 +174,7 @@ void buffer_set_slice(Buffer buffer,
   }
 
   size_t old_len = (size_t)(end - start);
-  if (!(buf->b_ml.ml_line_count + (new_len - old_len)) <= LONG_MAX) {
+  if (!((size_t)buf->b_ml.ml_line_count + (new_len - old_len) <= LONG_MAX)) {
     set_api_error("Exceding maximum line number", err);
     return;
   }
